[MO] - [commend] -> solve;
